chore(judicial-system): rename helpers/ to svgs/ in formatters
