+++ apps/judicial-system/backend/src/app/formatters/svgs/PDFKitCoatOfArms.ts
