+++ apps/judicial-system/backend/src/app/formatters/svgs/coatOfArms.ts
