+++ apps/judicial-system/backend/src/app/formatters/svgs/policeStar.ts
